% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aic.cutover.R
\name{aic.cutover}
\alias{aic.cutover}
\title{The AIC for a given generalized linear model with log-logit link.}
\usage{
aic.cutover(cutover, formula, ..., stop = F)
}
\arguments{
\item{cutover}{The cutover.}

\item{formula}{The regression formula, see help(glm).}

\item{...}{Other arguments to pass to glm2 (e.g. data).}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The Akaike Information Criterion (AIC) for the model (a single numeric value).
}
\description{
The loglogit link function is a log function up to a specified cutover point, and a scaled logit function
above this point.
}
\examples{
aic.cutover(0.6,y~x1+x2 , data=loglogit.example)
aic.cutover(0.7,y~x1+x2 , data=loglogit.example)
aic.cutover(0.8,y~x1+x2 , data=loglogit.example)
aic.cutover(0.9,y~x1+x2 , data=loglogit.example)
}

