% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blended.link.R
\name{blended.linkobject}
\alias{blended.linkobject}
\title{Link function which equals log function between user-specified cutoffs, and a scaled
  user-specified link function for linear predictors below the first cutoff and above
  the second cutoff.}
\usage{
blended.linkobject(otherlinkname, cutover)
}
\arguments{
\item{otherlinkname}{Character string indicating the link function to be used below cutover[1]
or above cutover[2]}

\item{cutover}{The link function switches smoothly from otherlinkname to log at cutover[1] and back to
otherlinkname at cutover[2]}
}
\value{
An object of class "link-glm", with link name "blended".
}
\description{
This function produces a blended object for a specified cutover and link function
for use in binary regression using glm or glm2.
}
\details{
This function returns a link function which which equals log function for
values of the fitted probability between cutover[1] and cutover[2].
For values less than cutover[1] or greater than cutover[2], a linearly scaled user-specified link
is used (typically logit is recommended). The linear scaling is such that the new link is smooth.
This means that the exponentiated regression coefficients can be interpreted as relative risks
for individuals whose probability is between cutover[1] and cutover[2] both with and without
the risk factor. In practice this is usually the great majority of cases.
The function is not normally called directly. Instead, glm or glm2 should be
used (preferably the latter to avoid fitting problems), as per the example below.
}
\examples{
library(glm2)
example.binary.regression <- glm2(y~x1+x2 , data=loglogit.example,
    family = binomial(link=blended.linkobject(cutover=c(0.2,0.8),otherlinkname="logit")))
summary(example.binary.regression)
}

