% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogit.dmu.deta.R
\name{loglogit.dmu.deta}
\alias{loglogit.dmu.deta}
\title{The derivative of the inverse function of the loglogit link function.}
\usage{
loglogit.dmu.deta(eta, cutover, stop = F)
}
\arguments{
\item{eta}{The linear predictors, a numeric vector of real values.}

\item{cutover}{The cutover.}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The derivatives of the link function calculated at the supplied linear predictor values.
}
\description{
The loglogit link function is a log function up to a specified cutover point, and a scaled logit function
above this point.
}
\examples{
loglogit.dmu.deta(eta=c(-2,-1,0,1,2),cutover=0.6)
}

