% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{loglogit.example}
\alias{loglogit.example}
\title{An example data file.}
\format{An object of class \code{data.frame} with 500 rows and 3 columns.}
\usage{
loglogit.example
}
\description{
A dataset with 500 rows containing the values of y (a binary dependent variable equal to 0 or 1),
and two covariates x1 and x2 (with continuous values ranging from -1 to 1).
}
\keyword{datasets}

