% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogit.linkfun.R
\name{loglogit.linkfun}
\alias{loglogit.linkfun}
\title{The loglogit link function.}
\usage{
loglogit.linkfun(mu, cutover, stop = F)
}
\arguments{
\item{mu}{The expected value, lying between 0 and 1.}

\item{cutover}{The cutover.}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The linear predictor corresponding to mu.
}
\description{
The loglogit link function is a log function up to a specified cutover point, and a scaled logit function
above this point.
}
\examples{
loglogit.linkfun(mu=c(0,0.25,0.5,0.75,1),cutover=0.6)
}

