% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogit.linkinv.R
\name{loglogit.linkinv}
\alias{loglogit.linkinv}
\title{The inverse function of the loglogit link function.}
\usage{
loglogit.linkinv(eta, cutover, stop = F)
}
\arguments{
\item{eta}{The linear predictors, a numeric vector of real values.}

\item{cutover}{The cutover.}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The expected values corresponding to the linear predictors.
}
\description{
The loglogit link function is a log function up to a specified cutover point, and a scaled logit function
above this point.
}
\examples{
loglogit.linkinv(eta=c(-2,-1,0,1,2),cutover=0.6)
}

