% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogit.link.R
\name{loglogit.linkobject}
\alias{loglogit.linkobject}
\title{Link function which equals log function up to a cutover point then a scaled logit function}
\usage{
loglogit.linkobject(cutover, stop = F)
}
\arguments{
\item{cutover}{The value at which the link function switches smoothly from log to scaled logit.}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
An object of class "link-glm", with link name "loglogit".
}
\description{
This function produces a blended log-logist link object for a specified cutover
for use in binary regression using glm or glm2.
}
\details{
This function returns a link function which which equals log function for
values less than or equal to a specified cutover point and a scaled logit
function for values above the cutover. This means that the
exponentiated regression coefficients can be interpreted as relative risks
for individuals whose probability is below the cutover both with and without
the risk factor. In practice this is usually the great majority of cases.
The function is not normally called directly. Instead, glm or glm2 should be
used (preferably the latter to avoid fitting problems), as per the example below.
}
\examples{
library(glm2)
example.loglogit.regression <- glm2(y~x1+x2 , data=loglogit.example,
    family = binomial(link=loglogit.linkobject(0.8)))
summary(example.loglogit.regression)
}

