% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglogit.valideta.R
\name{loglogit.valideta}
\alias{loglogit.valideta}
\title{The valideta item of the loglogit link object}
\usage{
loglogit.valideta(eta)
}
\arguments{
\item{eta}{The linear predictors, a numeric vector of real values.}
}
\value{
a logical vector of the same length as eta, with all values equal to TRUE
}
\description{
For internal use only.
}

