% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.cutover.R
\name{optim.cutover}
\alias{optim.cutover}
\title{Maximum likelihood estimator (MLE) of the cutover}
\usage{
optim.cutover(formula, data, ..., stop = F)
}
\arguments{
\item{formula}{The regression formula, see help(glm).}

\item{data}{A data frame containing the data to be analysed.}

\item{...}{additional named arguments to be passed to optimize}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The MLE of the cutover.
}
\description{
The value of the cutover which maximises the likelihood of the GLM
with the log-logit link for a given regression formula and dataset
}
\details{
This function is primarily for internal use and would not normally be used directly.
In general the cutover should be a large value (e.g. 0.6-0.8) so that exponentiated coefficients
can be interpreted as odds ratios for the majority of subjects. The cutover value should be
consistent with the data but not the MLE.
}
\examples{
optim.cutover(y~x1+x2 , data=loglogit.example)
}

