% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.cutover.MOM.R
\name{pvalue.cutover.MOM}
\alias{pvalue.cutover.MOM}
\title{Test null hypothesis that cutover is equal to a particular value}
\usage{
pvalue.cutover.MOM(cutover, formula, B = 30, data, stop = F)
}
\arguments{
\item{cutover}{The hypothesised value of the cutover}

\item{formula}{The regression formula, see help(glm).}

\item{B}{Number of boostraps to estimate the first and second moments of the
likelihood ratio test statistic.}

\item{data}{Data frame.}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The p-value for the hypothesis test.
}
\description{
Tests the null hypothesis that the cutover is equal to a supplied value for a given dataset
and regression formula
}
\details{
The p-value is based on the asymptotic chi-square distribution of 
the likelihood ratio test, and is not reliable for hypothesized cutovers
near the boundaries of the parameter space (i.e. close to 0 or 1)
}
\examples{
pvalue.cutover(cutover=0.8,y~x1+x2 , data=loglogit.example )
}

