% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.cutover.R
\name{pvalue.cutover}
\alias{pvalue.cutover}
\title{Test null hypothesis that cutover is equal to a particular value}
\usage{
pvalue.cutover(cutover, formula, ..., stop = F)
}
\arguments{
\item{cutover}{The hypothesised value of the cutover}

\item{formula}{The regression formula, see help(glm).}

\item{...}{Arguments to pass to optim.cutover and glm2, would typically consist of data, 
a data frame containing the data to be analysed.}

\item{stop}{Set to TRUE to open a browser window for debugging.}
}
\value{
The p-value for the hypothesis test.
}
\description{
Tests the null hypothesis that the cutover is equal to a supplied value for a given dataset
and regression formula
}
\details{
The p-value is based on the asymptotic chi-square distribution of 
the likelihood ratio test, and is not reliable for hypothesized cutovers
near the boundaries of the parameter space (i.e. close to 0 or 1)
}
\examples{
pvalue.cutover(cutover=0.8,y~x1+x2 , data=loglogit.example )
}

