% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalue.cutover.R
\name{pvalue.cutover}
\alias{pvalue.cutover}
\title{p-value of hypothesis test that the cutover of a blended link is equal to
a specific value}
\usage{
pvalue.cutover(data, formula, link1, link2, cutover0)
}
\arguments{
\item{link1}{Character string indicating the link function to be used up to the cutover}

\item{link2}{Character string indicating the link function to be used above the cutover}

\item{cutover0}{The hypothesized value of the cutover at which the link
function switches smoothly from link1 to link2. Must be strictly greater 
than 0 and strictly less than 1.}
}
\value{
The p-value of a test of the null hypothesis that cutover=cutover0.
}
\description{
This function calculates an asymptotic chi-square likelihood ratio
hypothesis test of a specified null hypothesis for the cutover.
}
\details{
Binary regression is assumed. Not suitable when the null value is on
 the boundary (i.e. when cutover0 is 0 or 1).
}
\examples{
library(glm2)
pvalue.cutover(y~x1+x2 , data=loglogit.example,link1="log",link2="logit",
cutover0=0.8)
}

